changed class name and stored strings into another file.
