+++ Assignment 2/Assignment-2/Constants.swift
@@ -1,0 +1,17 @@
+//
+//  Constants.swift
+//  Assignment-2
+//
+//  Created by Shubham on 10/02/22.
+//
+
+import Foundation
+
+struct Constants{
+    static let subjectA = "A"
+    static let subjectB = "B"
+    static let subjectC = "C"
+    static let subjectD = "D"
+    static let subjectE = "E"
+    static let subjectF = "F"
+}
